+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### 0.0.6-dev-pre
 
+- Autosave articles
+- Now non-writers cannot access the write page
+
 ### 0.0.5-dev
 
 - implement article backend